Move MeridionalHeatDiffusion to diffusion module where it belongs
