+++ climlab/dynamics/__init__.py
@@ -1,4 +1,4 @@
 '''Modules for horizontal transport processes in climlab.'''
 from __future__ import absolute_import
 from .budyko_transport import BudykoTransport
-from .diffusion import Diffusion, MeridionalDiffusion
+from .diffusion import Diffusion, MeridionalDiffusion, MeridionalHeatDiffusion
