+++ climlab/dynamics/diffusion.py
@@ -159,7 +159,7 @@ class Diffusion(ImplicitProcess):
                 self.diffusive_flux[1:-1] = -K[1:-1] * np.diff(np.squeeze(value), axis=ax)/self.delta  # length / time
                 self.diffusive_flux_convergence[:] = -np.diff(self.diffusive_flux*self._weight1, axis=ax)/self.delta/self._weight2 # 1/time
             else:
-                pass  # To do: implement flux diagnostics for arbitrary dimension state variables 
+                pass  # To do: implement flux diagnostics for arbitrary dimension state variables
 
 def _solve_implicit_banded(current, banded_matrix):
     """Uses a banded solver for matrix inversion of a tridiagonal matrix.
@@ -250,6 +250,53 @@ class MeridionalDiffusion(Diffusion):
             latax = dom.axes['lat']
         self.diffTriDiag, self._weight1, self._weight2 = \
             _make_meridional_diffusion_matrix(self.K_dimensionless, latax)
+
+
+class MeridionalHeatDiffusion(MeridionalDiffusion):
+    '''A 1D diffusion solver for Energy Balance Models.
+
+    Solves the meridional heat diffusion equation
+
+    $$ C \frac{\partial T}{\partial t} = -\frac{1}{\cos\phi} \frac{\partial}{\partial \phi} \left[ -D \cos\phi \frac{\partial T}{\partial \phi} \right]$$
+
+    on an evenly-spaced latitude grid, with a state variable $T$, a heat capacity $C$ and diffusivity $D$.
+
+    Assuming $T$ is a temperature in $K$ or $^\circ$C, then the units are:
+
+    - $D$ in W m$^{-2}$ K$^{-1}$
+    - $C$ in J m$^{-2}$ K$^{-1}$
+
+    If the state variable has other units, then $D$ and $C$ should be expressed
+    per state variabe unit.
+
+    $D$ is provided as input, and can be either scalar
+    or vector defined at latitude boundaries (length).
+
+    $C$ is normally handled automatically for temperature state variables in CLIMLAB.
+    '''
+    def __init__(self,
+                 D=0.555,  # in W / m^2 / degC
+                 use_banded_solver=True,
+                 **kwargs):
+        self.D = D
+        state = kwargs['state']
+        for varname, value in state.items():
+            heat_capacity = value.domain.heat_capacity
+        # diffusivity in units of 1/s
+        K = self.D / heat_capacity
+        super(MeridionalHeatDiffusion, self).__init__(K=K,
+                        use_banded_solver=use_banded_solver, **kwargs)
+        self.add_diagnostic('heat_transport', np.zeros_like(self.lat_bounds))
+        self.add_diagnostic('heat_transport_convergence', np.zeros_like(self.lat))
+
+    def _update_diagnostics(self, newstate):
+        super(MeridionalHeatDiffusion, self)._update_diagnostics(newstate)
+        for varname, value in self.state.items():
+            heat_capacity = value.domain.heat_capacity
+        self.heat_transport[:] = (self.diffusive_flux * heat_capacity *
+                        2 * np.pi * const.a * self._weight1 * 1E-15) # in PW
+        self.heat_transport_convergence[:] = (self.diffusive_flux_convergence *
+                        heat_capacity)  # in W/m**2
 
 
 def _make_diffusion_matrix(K, weight1=None, weight2=None):
