+++ conda-recipe/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "climlab" %}
-{% set version = "0.6.6.dev3" %}
+{% set version = "0.6.6.dev4" %}
 
 package:
   name: {{ name|lower }}
